test: #248 snapshot sibling S3 state across scoped init rerun
The NewObjects key diff misses in-place overwrites (deterministic keys)
and buildBasePaths misses same-path manifest rewrites. Snapshot the
sibling partition + manifest (size/ETag/LastModified + raw bytes) before
the target-scoped rerun and require exact identity after. Hoists
snapshotS3Inventory/snapshotManifest into a shared helpers file now that
they have a second consumer.
